parameter acceptance and input file safety

+++ dailyRun.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
+#make sure have parameter
+if [[ $# == 0 ]]; then
+    echo "missing required parameter"
+    exit 0
+fi
+
 #times run frontend per day
-RUNS=3
+RUNS=$1
 
 #clear out any old runs
 rm actualoutput/* 2> /dev/null #if nothing to remove then no errors
@@ -18,16 +24,23 @@ if [[ ! -t 0 ]]; then
     #loop through input file and separate it into temporary files for each run of the frontend
     num=1
     while read -r line; do
-        echo "$line yep"
         echo "$line" >> tempfile$num
         if [[ $line == "logout" ]]; then
             ((num+=1))
         fi
     done
 
+    #putting num back to the number of files that exist
+    ((num-=1))
+    if [[ $num != $RUNS ]]; then
+        echo "input file does not have an equal number of runs as requested daily runs"
+        rm tempfile* #clean up incomplete run
+        exit 0
+    fi
+
     #run the frontend for the day
     for (( i = 1; i <= $RUNS; i++ )); do
-        python3 code/frontEnd.py serviceList.txt actualoutput/${i}.txt < tempfile$i
+        python3 code/frontEnd.py serviceList.txt actualoutput/${i}.txt < tempfile$i > /dev/null #if automated input do not see prompts
         rm tempfile$i
     done
 
@@ -35,7 +48,7 @@ if [[ ! -t 0 ]]; then
 else
     #run the frontend for the day with user input
     for (( i = 1; i <= $RUNS; i++ )); do
-        python3 code/frontEnd.py serviceList.txt actualoutput/${i}.txt < file4
+        python3 code/frontEnd.py serviceList.txt actualoutput/${i}.txt
     done
     
 fi
